Refactor create-slideshow function
Refactor the create-slideshow Supabase function into smaller, more manageable files for improved readability and maintainability.
