+++ supabase/functions/create-slideshow/utils/formatter.ts
@@ -1,0 +1,9 @@
+
+export const formatPrice = (price: number) => {
+  return new Intl.NumberFormat("fr-CA", {
+    style: "currency",
+    currency: "CAD",
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 0,
+  }).format(price);
+};
