+++ supabase/functions/create-slideshow/utils/textElements.ts
@@ -1,0 +1,33 @@
+
+import { formatPrice } from "./formatter.ts";
+
+export const prepareTextElements = (listing: any, config: any) => {
+  const textElements = [];
+  
+  if (config.showDetails) {
+    if (config.showPrice && listing.price) {
+      textElements.push(formatPrice(listing.price));
+    }
+    
+    if (config.showAddress && listing.address) {
+      let address = listing.address;
+      if (listing.city) {
+        address += `, ${listing.city}`;
+      }
+      if (listing.postal_code) {
+        address += ` ${listing.postal_code}`;
+      }
+      textElements.push(address);
+    }
+    
+    const details = [];
+    if (listing.bedrooms) details.push(`${listing.bedrooms} ch.`);
+    if (listing.bathrooms) details.push(`${listing.bathrooms} sdb.`);
+    if (listing.property_type) details.push(listing.property_type);
+    if (details.length > 0) {
+      textElements.push(details.join(" | "));
+    }
+  }
+  
+  return textElements;
+};
